Move block-related type aliases into a separate module
