+++ subspace/crates/sc-consensus-subspace/src/slot_worker.rs
@@ -47,7 +47,7 @@ use std::marker::PhantomData;
 use std::pin::Pin;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
-use subspace_core_primitives::BlockNumber;
+use subspace_core_primitives::block::BlockNumber;
 use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::pot::{PotCheckpoints, PotOutput, SlotNumber};
 use subspace_core_primitives::sectors::SectorId;
