+++ subspace/crates/subspace-archiving/src/reconstructor.rs
@@ -3,7 +3,7 @@ use ab_erasure_coding::{ErasureCoding, ErasureCodingError, RecoveryShardState};
 use alloc::vec::Vec;
 use core::mem;
 use parity_scale_codec::Decode;
-use subspace_core_primitives::BlockNumber;
+use subspace_core_primitives::block::BlockNumber;
 use subspace_core_primitives::pieces::Piece;
 use subspace_core_primitives::segments::{
     ArchivedBlockProgress, ArchivedHistorySegment, LastArchivedBlock, RecordedHistorySegment,
