+++ subspace/crates/subspace-service/src/sync_from_dsn/import_blocks.rs
@@ -14,7 +14,7 @@ use sp_runtime::traits::{Block as BlockT, Header, NumberFor};
 use std::sync::{Arc, Mutex};
 use std::time::Duration;
 use subspace_archiving::reconstructor::Reconstructor;
-use subspace_core_primitives::BlockNumber;
+use subspace_core_primitives::block::BlockNumber;
 use subspace_core_primitives::segments::SegmentIndex;
 use subspace_data_retrieval::segment_downloading::download_segment_pieces;
 use tokio::task;
