+++ subspace/crates/subspace-service/src/utils.rs
@@ -3,7 +3,7 @@ use sc_client_api::BlockchainEvents;
 use sp_blockchain::HeaderBackend;
 use sp_runtime::SaturatedConversion;
 use sp_runtime::traits::{Block as BlockT, Header};
-use subspace_core_primitives::BlockNumber;
+use subspace_core_primitives::block::BlockNumber;
 use tracing::{debug, trace};
 
 pub async fn wait_for_block_import<Block, Client>(
