+++ subspace/crates/subspace-verification/src/lib.rs
@@ -8,7 +8,7 @@ pub mod sr25519;
 use schnorrkel::SignatureError;
 use schnorrkel::context::SigningContext;
 use sr25519::RewardSignature;
-use subspace_core_primitives::BlockWeight;
+use subspace_core_primitives::block::BlockWeight;
 use subspace_core_primitives::hashes::Blake3Hash;
 use subspace_core_primitives::solutions::SolutionRange;
